+++ test/test.py
@@ -21,9 +21,16 @@ def setup():
     lib_dir = join(top_dir, "lib")
     sys.path.insert(0, lib_dir)
 
-    pygments_dir = join(top_dir, "deps", "pygments")
-    if exists(pygments_dir):
-        sys.path.insert(0, pygments_dir)
+    # Attempt to get 'pygments' on the import path.
+    try:
+        # If already have it, use that one.
+        import pygments
+    except ImportError:
+        pygments_dir = join(top_dir, "deps", "pygments")
+        if sys.version_info[0] <= 2:
+            sys.path.insert(0, pygments_dir)
+        else:
+            sys.path.insert(0, pygments_dir + "3")
 
 if __name__ == "__main__":
     logging.basicConfig()