+++ test/testall.py
@@ -24,7 +24,8 @@ def _python_ver_from_python(python):
 
 def _gen_python_names():
     yield "python"
-    for ver in [(2,4), (2,5), (2,6), (2,7), (3,0), (3,1), (3,2)]:
+    for ver in [(2,2), (2,3), (2,4), (2,5), (2,6), (2,7), (3,0), (3,1),
+                (3,2), (3,3)]:
         yield "python%d.%d" % ver
         if sys.platform == "win32":
             yield "python%d%d" % ver
@@ -43,9 +44,6 @@ def testall():
     for ver, python in _gen_pythons():
         if ver < (2,3):
             # Don't support Python < 2.3.
-            continue
-        elif ver >= (3, 0):
-            # Don't yet support Python 3.
             continue
         ver_str = "%s.%s" % ver
         print "-- test with Python %s (%s)" % (ver_str, python)