+++ CHANGES.md
@@ -1,8 +1,9 @@
 # python-markdown2 Changelog
 
-## python-markdown2 1.3.2 (not yet released)
+## python-markdown2 1.4.0 (not yet released)
 
-(nothing yet)
+- [issue #64] Python 3 support! markdown2.py supports Python 2 and 3 in the
+  same file without requiring install-time 2to3 transformation.
 
 
 ## python-markdown2 1.3.1